Odida Jared.md
I have added a few updates on the buyer registration content, eliminating and proposing additional statements on some sections. This is to make it easier for understanding by the client making the registration. However, more needs to be done to make it more clear as some may not be well understood.

Thank you